% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfclim-package.R
\docType{package}
\name{mfclim-package}
\alias{mfclim}
\alias{mfclim-package}
\title{mfclim: Download Archived Meteorological Data from 'Meteo-France'}
\description{
Download meteorological data from the 'Meteo-France' public API 'Données climatologiques' \url{https://portail-api.meteofrance.fr} and 'SYNOP' open data archives \url{https://meteo.data.gouv.fr/datasets/686f8595b351c06a3a790867}. It provides functions to authenticate, list stations, retrieve station metadata, request data, download files, and import data directly into 'R'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Nmoiroux/mfclim}
  \item Report bugs at \url{https://github.com/Nmoiroux/mfclim/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Moiroux \email{nicolas.moiroux@ird.fr} (\href{https://orcid.org/0000-0001-6755-6167}{ORCID})

}
\keyword{internal}
