% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{mfclim_download}
\alias{mfclim_download}
\title{Download a climate data file from the Meteo-France API}
\usage{
mfclim_download(token, id_cmde, filename = "data.csv", folder = tempdir())
}
\arguments{
\item{token}{Character string. OAuth2 access token obtained with
\code{mfclim_get_token()}.}

\item{id_cmde}{Numeric or character. Order ID returned by
\code{mfclim_order()}.}

\item{filename}{Character string. Name of the output CSV file.}

\item{folder}{Character string. Where to save the downloaded CSV file. Default to \code{tempdir()}}
}
\value{
Character string. Path to the downloaded file.
}
\description{
Download the CSV file corresponding to a previously submitted data order
. The function repeatedly queries the API until the file is ready
and then downloads it to disk.
}
\details{
The function polls the API every 5 seconds until the file is
available. When the file is ready, it is downloaded and saved to
\code{filename}.
}
\examples{
\dontrun{
file <- mfclim_download(token, id_cmde = 123456, filename = "meteo.csv")
}
}
\seealso{
\code{mfclim_order()}
}
