% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{mfclim_parse_json}
\alias{mfclim_parse_json}
\title{Parse a JSON-like list from Météo-France API into R structures}
\usage{
mfclim_parse_json(x)
}
\arguments{
\item{x}{A named list, typically obtained from \code{httr2::content(..., as="parsed")}.}
}
\value{
A named list where:
\itemize{
\item simple values are stored as vectors,
\item nested lists are converted into data.frames via \code{do.call(rbind, ...)}.
}
}
\description{
Converts a parsed JSON object (list) returned by the Météo-France API
into a more convenient R structure. Simple values are stored as vectors,
while nested lists are converted into data.frames.
}
\details{
Used for transforming the structure returned by
\code{mfclim_info_station} into a more user-friendly format.
}
