% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{mfclim_synop}
\alias{mfclim_synop}
\title{Download SYNOP data from Meteo-France open data}
\usage{
mfclim_synop(year, path = tempdir())
}
\arguments{
\item{year}{Numeric or character. Year of the data to download (YYYY), between 1996 and present year (included).}

\item{path}{Path to folder where to save the archive. Default to \code{tempdir()}.}
}
\value{
A data.frame containing the downloaded data.
}
\description{
Downloads SYNOP surface observation data from the Meteo-France open data
portal for a given year, decompresses the archive, and return a data frame.
}
\details{
SYNOP data come from observation messages exchanged
through the World Meteorological Organization (WMO) Global Telecommunication
System (GTS). These observations include measured atmospheric parameters
such as temperature, humidity, wind direction and speed, atmospheric pressure,
and precipitation amount, as well as observed variables such as present weather,
cloud description, and visibility. Depending on instrumentation and local
conditions, additional variables may be available (e.g., snow depth, ground
state).

The dataset covers mainland France and overseas territories with a typical
observation frequency of 3 hours. Available years cover 1996 to present.

Data description and documentation is available here (French):
\url{https://meteo.data.gouv.fr/datasets/686f8595b351c06a3a790867}

Files are downloaded as \code{.csv.gz}, decompressed on the fly, and read.
Typical size of compressed files is <7MB for period 1996-2024 and <20MB for more recent years.
Missing values coded as "", "NA", or "mq" are converted to NA in the returned data frame.
}
\examples{
\dontrun{
mydata <- mfclim_synop(2020)
}
}
